build: upgrade asp-net example to version 9.0
Refs: #28
