+++ AspNetAuthExample/AspNetAuthExample.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.12.35728.132 d17.12
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AspNetAuthExample", "AspNetAuthExample.csproj", "{208859B8-8540-43F1-AE46-902336890B9B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{208859B8-8540-43F1-AE46-902336890B9B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{208859B8-8540-43F1-AE46-902336890B9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{208859B8-8540-43F1-AE46-902336890B9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{208859B8-8540-43F1-AE46-902336890B9B}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
